Fix the gas station to work with float values

+++ clients/gasManagement/export_test.go
@@ -3,7 +3,7 @@ package gasManagement
 import "github.com/multiversx/mx-bridge-eth-go/core"
 
 // GetLatestGasPrice -
-func (gs *gasStation) GetLatestGasPrice() int {
+func (gs *gasStation) GetLatestGasPrice() float64 {
 	gs.mut.RLock()
 	defer gs.mut.RUnlock()
